feat: add change pages
