+++ src/index.html
@@ -1,7 +1,6 @@
 <!doctype html>
 <html lang="ru">
 
-
 <head>
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -14,4 +13,3 @@
 </body>
 
 </html>
-
