+++ src/layout/create.scss
@@ -1,0 +1,96 @@
+.create-page {
+  padding: 40px 0px;
+  // .create-page__title
+  &__title {
+  }
+  // .create-page__body
+  &__body {
+  }
+  // .create-page__content
+  &__content {
+    display: flex;
+    justify-content: space-between;
+    margin: 0px 0px 50px 0px;
+  }
+  // .create-page__view
+  &__view {
+  }
+  // .create-page__chart
+  &__chart {
+  }
+  // .create-page__actions
+  &__actions {
+  }
+}
+
+.how-create {
+  // .how-create__body
+  &__body {
+    display: flex;
+    gap: 10px;
+    justify-content: space-between;
+  }
+  // .how-create__title
+  &__title {
+    font-size: 25px;
+  }
+  // .how-create__description
+  &__description {
+  }
+  // .how-create__step
+  &__step {
+    font-size: 18px;
+    span {
+      font-weight: 700;
+    }
+  }
+  &__buttons {
+  }
+  // .how-create__read
+  &__read {
+  }
+  // .how-create__look
+  &__look {
+  }
+}
+.one {
+  color: $blue;
+}
+.two {
+  color: $red;
+}
+.three {
+  color: $yellow;
+}
+.button-more {
+  display: inline-flex;
+  align-items: center;
+  gap: 10px;
+  padding: 8px 16px;
+  font-size: 18px;
+  font-weight: 600;
+  border: 2px solid #9e9898;
+  border-radius: 20px;
+  background-color: #fff;
+  &:not(:last-child) {
+    margin: 0px 15px 0px 0px;
+  }
+  transition: all 0.3s;
+  &::after {
+    content: "";
+    background: url("../assets/arrow.svg") no-repeat center / cover;
+    background-size: contain;
+    width: 12px;
+    height: 12px;
+  }
+  @media (any-hover: hover) {
+    &:hover {
+      color: $blue;
+      border: 2px solid $blue;
+      background-color: #fff;
+      &::after {
+        transform: scale(1.3);
+      }
+    }
+  }
+}
