+++ src/layout/header.scss
@@ -1,0 +1,112 @@
+.header {
+  // .header__body
+  &__body {
+    display: flex;
+    align-items: center;
+    gap: 30px;
+  }
+
+  // .header__logo
+  &__logo {
+    img {
+      width: 100px;
+      height: 80px;
+      object-fit: contain;
+    }
+  }
+  // .header__content
+  &__content {
+    margin: 0px auto 0px 0px;
+  }
+  // .header__title
+  &__title {
+    display: inline-block;
+    font-size: 18px;
+    font-weight: 700;
+    padding: 10px 0px 10px 0px;
+    transition: all 0.3s;
+    position: relative;
+    color: #000000;
+    &::before {
+      position: absolute;
+      top: 0;
+      left: 0;
+      content: "";
+      width: 100%;
+      height: 2px;
+    }
+    @media (any-hover: hover) {
+      &:hover {
+        color: $blue;
+        &::before {
+          background-color: $blue;
+        }
+      }
+    }
+  }
+}
+
+.menu {
+  // .menu__list
+  &__list {
+    display: flex;
+    gap: 20px;
+  }
+  // .menu__link
+  &__link {
+    font-size: 16px;
+    font-weight: 600;
+    transition: all 0.3s;
+    padding: 0px 0px 10px 0px;
+    border-bottom: 2px solid #ffffff00;
+    color: #9e9898;
+    @media (any-hover: hover) {
+      &:hover {
+        border-bottom: 2px solid $yellow;
+        color: $yellow;
+      }
+    }
+  }
+}
+
+.actions {
+  display: flex;
+  align-items: center;
+  gap: 25px;
+  &__bg {
+    width: 30px;
+    height: 30px;
+    transition: all 0.3s;
+    &_black {
+      background: url("../assets/moon.svg") no-repeat center / cover;
+    }
+    &.white {
+      background: url("../assets/sun.svg") no-repeat center / cover;
+    }
+    @media (any-hover: hover) {
+      &:hover {
+        transform: scale(1.2);
+      }
+    }
+  }
+}
+.actions__language {
+  transition: all 0.3s;
+  &.ru {
+    background: url("../public/russia.png");
+    background-size: contain;
+    width: 30px;
+    height: 30px;
+  }
+  &.uk {
+    background: url("../public/UK.png");
+    background-size: contain;
+    width: 30px;
+    height: 30px;
+  }
+  @media (any-hover: hover) {
+    &:hover {
+      transform: scale(1.2);
+    }
+  }
+}
