+++ src/layout/main.scss
@@ -1,8 +1,3 @@
-$red: rgb(226, 86, 89);
-$yellow: rgb(242, 143, 39);
-$blue: rgb(77, 122, 169);
-$green: rgb(119, 184, 180);
-
 .main-page {
   margin: 0 auto;
   text-align: center;
@@ -23,9 +18,10 @@ $green: rgb(119, 184, 180);
   }
 
   &__h2 {
-    margin: 10px 0 30px;   
+    margin: 10px 0 30px;
     color: #bbb;
     font-weight: 300;
+    font-size: 25px;
   }
 
   &__button {
@@ -39,19 +35,21 @@ $green: rgb(119, 184, 180);
 
   &__img {
     width: 50px;
+    margin: 0px 0px 20px 0px;
   }
 
   &__gallery {
     background-color: #fff;
     padding: 20px 0;
     border-radius: 10px;
-    box-shadow: 0 -1px 2px -1px rgb(0 0 0 / 30%), 0 1px 2px -1px rgb(0 0 0 / 40%);
+    box-shadow: 0 0px 0px 1px rgb(242 233 233 / 30%), 0 0px 0px 0px rgb(249 249 249 / 99%);
     border: 1px solid transparent;
   }
 
   &__gallery-h2 {
     margin-bottom: 20px;
     font-weight: normal;
+    font-size: 25px;
   }
 
   &__text {
@@ -62,7 +60,6 @@ $green: rgb(119, 184, 180);
 
 .img-block {
   width: 140px;
-  font-size: 14px;
 }
 
 .animation__chart-column {
