+++ src/layout/popupLanguage.scss
@@ -1,0 +1,96 @@
+.popup {
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  z-index: 3;
+  &::before {
+    content: "";
+    background-color: rgba(233, 232, 232, 0.6);
+    backdrop-filter: blur(5px);
+    position: fixed;
+    width: 100%;
+    height: 100%;
+    top: 0;
+  }
+  // .popup__content
+  &__content {
+    padding: 25px 70px;
+    border-radius: 15px;
+    border: 1px solid #9e9898;
+    position: relative;
+    background-color: #fff;
+    z-index: 4;
+  }
+  // .popup__close
+  &__close {
+    position: absolute;
+    top: 15px;
+    right: 20px;
+    z-index: 4;
+    &::before {
+      content: "";
+      position: absolute;
+      top: 0px;
+      left: -10px;
+      width: 20px;
+      height: 3px;
+      background-color: #000000;
+      transform: rotate(135deg);
+    }
+    &::after {
+      content: "";
+      position: absolute;
+      top: 0px;
+      left: -10px;
+      width: 20px;
+      height: 3px;
+      background-color: #000000;
+      transform: rotate(-135deg);
+    }
+  }
+
+  // .popup__body
+  &__body {
+    display: inline-flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+  // .popup__lang
+  &__lang {
+    font-weight: 700;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    transition: all 0.3s;
+    cursor: pointer;
+    &_ru {
+      &::before {
+        content: "";
+        background: url("../public/russia.png");
+        background-size: contain;
+        width: 25px;
+        height: 25px;
+      }
+    }
+    &_en {
+      &::before {
+        content: "";
+        background: url("../public/UK.png");
+        background-size: contain;
+        width: 25px;
+        height: 25px;
+      }
+    }
+    @media (any-hover: hover) {
+      &:hover {
+        color: $yellow;
+        text-decoration: underline;
+      }
+    }
+  }
+}
