+++ src/style.scss
@@ -1,4 +1,3 @@
-
 $fontFamily: "";
 $fontSize: 16;
 
@@ -15,7 +14,9 @@ $blue: rgb(77, 122, 169);
 $green: rgb(119, 184, 180);
 
 @import "./utils/null.scss";
-@import "./page-create/page-create.scss";
+@import "./layout/header.scss";
+@import "./layout/popupLanguage.scss";
+@import "./layout/create.scss";
 @import "./layout/info.scss";
 @import "./layout/main.scss";
 @import "./layout/gallery.scss";
@@ -31,6 +32,12 @@ html {
     .settings-actions,
     .btn-add,
     .popup__content,
+    .menu__link,
+    .gallery-block,
+    .info-page,
+    .info-page__section,
+    .main-page__start,
+    .main-page__gallery,
     a {
       background-color: #000000;
       color: #ffffff;
@@ -50,7 +57,7 @@ html {
     }
   }
 }
-=======
+
 .wrapper {
   width: 100%;
   min-height: 100%;
@@ -59,9 +66,6 @@ html {
   flex-direction: column;
   position: relative;
 }
-
-
-
 .page {
   flex: 1 1 auto;
   background: #fafafa;
@@ -72,4 +76,9 @@ html {
   max-width: $maxWidthContainer + px;
   margin: 0 auto;
   padding: 0 15px;
+}
+
+.main-title {
+  font-size: 31px;
+  margin: 0px 0px 15px 0px;
 }
