+++ src/utils/translate.ts
@@ -29,11 +29,11 @@ export const lang:LangInt = {
     'ru':'Готово!',
     'uk':' Done!',
   },
-  'btncreate1':{
+  'btnmore1':{
     'ru':'Читать подробнее',
     'uk':'Read more',
   },
-  'btncreate2':{
+  'btnmore2':{
     'ru':'Смотреть примеры',
     'uk':'See examples',
   },
@@ -48,5 +48,132 @@ export const lang:LangInt = {
   'how-create__title':{
     'ru':'Как создать график/диаграмму онлайн',
     'uk':'How to create a chart/diagram online on the website',
+  },
+  'gallery-page__h1': {
+    'ru': 'Галерея графиков/диаграмм',
+    'uk': 'Chart/diagram gallery',
+  },
+  'gallery-page__text': {
+    'ru': `Выберите необходимый вид графика/диаграммы, далее вы перейдёте в режим 
+      онлайн-конструктора, в котором сможете заполнить данные графика, и затем 
+      скачать или сохранить вашу диаграмму. Какую диаграмму вам необходимо построить? 
+      Линейную, круговую, столбчатую, полосовую, линии, области, …?`,
+    'uk': `Choose the desired type of chart/diagram, then you will get to the online 
+      constructor page, where you can fill in the chart data, and then download or save your 
+      chart. What diagram do you need to build? Linear, pie, columnar, bar, lines, areas, …?`,
+  },
+  'gallery-block-name1': {
+    'ru': 'Колонны',
+    'uk': 'Columns',
+  },
+  'gallery-block-name2': {
+    'ru': 'Пончик',
+    'uk': 'Donut',
+  },
+  'gallery-block-name3': {
+    'ru': 'Полосы',
+    'uk': 'Bars',
+  },
+  'gallery-block-name4': {
+    'ru': 'Линии',
+    'uk': 'lines',
+  },
+  'gallery-block-name5': {
+    'ru': 'Пирог',
+    'uk': 'Pie',
+  },
+  'gallery-block-name6': {
+    'ru': 'Радар',
+    'uk': 'Radar',
+  },
+  'main-page__h2': {
+    'ru': 'Графики/диаграммы онлайн',
+    'uk': 'Online charts/diagrams',
+  },
+  'info-page__button': {
+    'ru': 'Создать график',
+    'uk': 'Create chart',
+  },
+  'img-block__text1': {
+    'ru': 'Создайте онлайн, бесплатно и без регистрации',
+    'uk': 'Create online, free and without registration',
+  },
+  'img-block__text2': {
+    'ru': 'Делитесь ссылками на ваши графики',
+    'uk': 'Share charts links',
+  },
+  'img-block__text3': {
+    'ru': 'Скачивайте графики',
+    'uk': 'Download charts',
+  },
+  'main-page__gallery-h2': {
+    'ru': 'Галерея графиков/диаграмм',
+    'uk': 'Chart/diagram gallery',
+  },
+  'main-page__text-p': {
+    'ru': `- сайт для построения графиков/диаграмм онлайн. Выберите вид диаграммы 
+      (линейная, круговая, столбчатая, полосовая, линии, области), заполните данные таблицы 
+      графика, и диаграмма готова! Диаграмму можно скачать в удобном вам формате или получить
+      ссылку на страницу вашего графика. Вы можете также использовать онлайн-конструктор графиков 
+      на нашем сайте, чтобы скачать график (png/jpeg/и др.) и вставить его в excel, word и другие 
+      программы на вашем компьютере. Построить диаграмму онлайн очень просто! Читайте подробнее о том, 
+      как создать диаграмму онлайн здесь. Также вы можете посмотреть примеры диаграмм, чтобы на их основе 
+      создать необходимый вам график.`,
+    'uk': `is a website for building charts/diagrams online. Select the chart type (linear, pie, columnar, bar, 
+      lines, areas), fill in the chart/diagram data table, and the chart is done! You can download the chart in a 
+      format that is convenient for you or get a link to the page of your chart/diagram. You can also use the online 
+      chart and diagram constructor on our website to download the chart (png / jpeg /, etc.) and paste it into 
+      Excel, Word, and other programs on your computer. It is very easy to create a chart/diagram online! Read more 
+      about how to create a diagram online here. Please see also the examples of the charts/diagrams to create the 
+      one you need.`,
+  },
+  'info-page__h1': {
+    'ru': 'Как создать график/диаграмму онлайн на сайте',
+    'uk': 'How to create a chart/diagram online on the website',
+  },
+  'info-page__text1': {
+    'ru': `Вы можете посмотреть примеры графиков/диаграмм и создать свой график из 
+      подходящего вам примера.`,
+    'uk': `You can find the chart/diagram examples and create your own one based on 
+      the example most suitable for you.`,
+  },
+  'info-page__text2': {
+    'ru': 'Ниже описана полная инструкция, как создать диаграммуна нашем сайте. Это просто:',
+    'uk': `Below you may find the full instruction on how to create the diagram on our website. 
+      It is very easy:`,
+  },
+  'section-view__h2': {
+    'ru': '1. Выберите вид графика',
+    'uk': '1. Select the chart type',
+  },
+  'section-view__text': {
+    'ru': `На странице создания или в галерее выберите подходящий вам тип графика.
+      Как вам необходимо отобразить ваши данные?
+      Линии / области / колонны / круговая диаграмма / … ?`,
+    'uk': `On the Creation page or in the Gallery, choose the chart/diagram type you need.
+      How do you need to display your data?
+      Lines / areas / columns / pie chart / … ?`,
+  },
+  'section-tables__h2': {
+    'ru': '2. Заполните таблицу данных',
+    'uk': '2. Fill in the data table',
+  },
+  'section-tables__text': {
+    'ru': `Для того, чтобы построить диаграмму, необходимо заполнить таблицу данных.
+      Значения в первой колонке станут отметками по горизонтали (по оси X).
+      Остальные колонки (их может быть одна или несколько) – это значения,
+      которые соответствуют отметкам из первой колонки, и которые будут отображены на диаграмме (ось Y).`,
+    'uk': `In order to create a chart, please fill in the data table.
+      The values in the first column will become the horizontal marks (along the X axis).
+      The remaining columns (there may be one or more) are the values
+      which correspond to the marks from the first column, and which will be displayed on the diagram (Y-axis).`,
+  },
+  'section-ready__h2': {
+    'ru': '3. Готово!',
+    'uk': '3. Done!',
+  },  
+  'section-ready__text': {
+    'ru': 'Теперь график можно скачать в удобном вам формате или поделиться ссылкой на вашу диаграмму.',
+    'uk': 'Now you can download the chart/diagram in a convenient format or share a link to your chart.',
   },
 };
